requirements.txt
beautifulsoup4==4.14.3
EbookLib==0.20
img2pdf==0.6.3
markdown-it-py==4.0.0
markdown-katex==202406.1035
openai==2.30.0
pillow==12.2.0
python-Levenshtein==0.27.3
PyMuPDF==1.27.2.2
pypdf==6.9.2
requests==2.33.1
tqdm==4.67.3
weasyprint==68.1
